embedding/requirements.txt
fastapi 
uvicorn[standard] 
transformers 
torch
librosa
pyannote.audio